[ENHANCEMENT] Add timerange and refresh buttons on the sticky toolbar
