+++ ui/dashboards/src/components/index.ts
@@ -15,6 +15,7 @@ export * from './AddGroupButton';
 export * from './AddPanelButton';
 export * from './Dashboard';
 export * from './DashboardToolbar';
+export * from './DashboardStickyToolbar';
 export * from './DeletePanelDialog';
 export * from './DeletePanelGroupDialog';
 export * from './DiscardChangesConfirmationDialog';
